changes to ui layout and seperated files to make easier to work with
