+++ lib/spotifyservice.dart
@@ -1,4 +1,5 @@
 import 'package:flutter_webview_plugin/flutter_webview_plugin.dart';
+
 class SpotifyService {
 
 }
